Add LaunchOnlyOnce key
<key>LaunchOnlyOnce</key>
<true/>

By adding this to the LaunchAgent then the window will actually quit and stay off. Not sure what the rationale is for having it relaunch every 10 seconds, particularly if you offer the option to quit the window, but if for some reason the removal of the .plist fails then the machine is practically unusable. By adding the LaunchOnlyOnce key then it will not respawn constantly.